requirements.txt
Flask-APScheduler==1.11.0
flask-redis==0.4.0
redis==3.3.11
az.cli